+++ sdk/core/core-client/src/serviceClient.ts
@@ -29,6 +29,7 @@ export interface ServiceClientOptions extends CommonClientOptions {
   /**
    * If specified, this is the base URI that requests will be made against for this ServiceClient.
    * If it is not specified, then all OperationSpecs must contain a baseUrl property.
+   * This property is deprecated and will be removed soon, please use endpoint instead
    * @deprecated
    */
   baseUri?: string;
@@ -70,7 +71,7 @@ export class ServiceClient {
    * If specified, this is the base URI that requests will be made against for this ServiceClient.
    * If it is not specified, then all OperationSpecs must contain a baseUrl property.
    */
-  private readonly _baseUri?: string;
+  private readonly _endpoint?: string;
 
   /**
    * The default request content type for the service.
@@ -100,7 +101,7 @@ export class ServiceClient {
    */
   constructor(options: ServiceClientOptions = {}) {
     this._requestContentType = options.requestContentType;
-    this._baseUri = options.baseUri;
+    this._endpoint = options.endpoint ?? options.baseUri;
     this._allowInsecureConnection = options.allowInsecureConnection;
     this._httpClient = options.httpClient || getCachedDefaultHttpClient();
 
@@ -134,8 +135,8 @@ export class ServiceClient {
     operationArguments: OperationArguments,
     operationSpec: OperationSpec
   ): Promise<T> {
-    const baseUri: string | undefined = operationSpec.baseUrl || this._baseUri;
-    if (!baseUri) {
+    const endpoint: string | undefined = operationSpec.baseUrl || this._endpoint;
+    if (!endpoint) {
       throw new Error(
         "If operationSpec.baseUrl is not specified, then the ServiceClient must have a baseUri string property that contains the base URL to use."
       );
@@ -144,7 +145,7 @@ export class ServiceClient {
     // Templatized URLs sometimes reference properties on the ServiceClient child class,
     // so we have to pass `this` below in order to search these properties if they're
     // not part of OperationArguments
-    const url = getRequestUrl(baseUri, operationSpec, operationArguments, this);
+    const url = getRequestUrl(endpoint, operationSpec, operationArguments, this);
 
     const request: OperationRequest = createPipelineRequest({
       url,
@@ -250,13 +251,13 @@ function getCredentialScopes(options: ServiceClientOptions): string | string[] |
       : new URL(scopes).toString();
   }
 
-  if (options.baseUri) {
-    return `${options.baseUri}/.default`;
+  if (options.endpoint) {
+    return `${options.endpoint}/.default`;
   }
 
   if (options.credential && !options.credentialScopes) {
     throw new Error(
-      `When using credentials, the ServiceClientOptions must contain either a baseUri or a credentialScopes. Unable to create a bearerTokenAuthenticationPolicy`
+      `When using credentials, the ServiceClientOptions must contain either a endpoint or a credentialScopes. Unable to create a bearerTokenAuthenticationPolicy`
     );
   }