+++ sdk/core/core-client/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added a new property endpoint in ServiceClientOptions and mark the baseUri as deprecated to encourage people to use endpoint.
+- Added a new property endpoint in ServiceClientOptions and mark the baseUri as deprecated to encourage people to use endpoint. See issue link [here](https://github.com/Azure/autorest.typescript/issues/1337)
 - Upgraded our `@azure/core-tracing` dependency to version 1.0
 
 ## 1.5.0 (2022-02-03)